update, get user and upload profile pic and change password

+++ src/main/java/tn/esprit/spring/Repository/IDepratmentRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.Repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.spring.Entity.Departement;
 
-public interface IDepratmentRepository extends CrudRepository<Departement, Integer> {
+public interface IDepratmentRepository extends CrudRepository<Departement, Long> {
 }
